+++ src/main/java/net/geforcemods/securitycraft/blockentities/KeypadChestBlockEntity.java
@@ -300,6 +300,14 @@ public class KeypadChestBlockEntity extends ChestBlockEntity implements IPasscod
 	}
 
 	@Override
+	public void setPasscodeInAdjacentBlock(String codeToSet) {
+		if (findOther() instanceof KeypadChestBlockEntity chestBe && getOwner().owns(chestBe)) {
+			chestBe.hashAndSetPasscode(codeToSet, getSalt());
+			level.sendBlockUpdated(chestBe.worldPosition, chestBe.getBlockState(), chestBe.getBlockState(), 2);
+		}
+	}
+
+	@Override
 	public Owner getOwner() {
 		return owner;
 	}