+++ Changelog.md
@@ -1,3 +1,7 @@
+--------------------------Changelog for v1.9.121 of SecurityCraft--------------------------
+
+- API: New method IPasscodeProtected#setPasscodeInAdjacentBlock for updating an adjacent block (e.g. the second half of a chest) with the passcode when setting it
+
 --------------------------Changelog for v1.9.11 of SecurityCraft--------------------------
 
 - New: More blocks now support the Disguise Module: Keypad Door, Keypad Trapdoor, Passcode-Protected Chest, Reinforced Dispenser, Reinforced Dropper, Reinforced Hopper, Scanner Door, Scanner Trapdoor, Security Camera, Sonic Security System